Move exceptions to public api folder
