+++ src/api/java/de/fearnixx/jeak/service/controller/exceptions/RegisterControllerException.java
@@ -1,4 +1,4 @@
-package de.fearnixx.jeak.service.controller;
+package de.fearnixx.jeak.service.controller.exceptions;
 
 public class RegisterControllerException extends RuntimeException{
     public RegisterControllerException(String message) {
